+++ fourier.cpp
@@ -11,8 +11,8 @@ std::vector<double> fourier(const std::vector<short> &samples) {
   auto *twiddle = new complex<double>[bins][bins]();
 
   // Populate twiddle matrix. The "exp" is the important bit.
-  for (unsigned int k = 0; k < bins; ++k)
-    for (unsigned int n = 0; n < bins; ++n)
+  for (unsigned long k = 0; k < bins; ++k)
+    for (unsigned long n = 0; n < bins; ++n)
       twiddle[n][k] =
           exp(complex<double>(0, 2) * M_PI * static_cast<double>(k) *
               static_cast<double>(n) / static_cast<double>(bins));
@@ -21,10 +21,10 @@ std::vector<double> fourier(const std::vector<short> &samples) {
   // original samples. Only run over the first half of the matrix as the other
   // half is a mirror image.
   vector<double> results;
-  for (unsigned int k = 0; k < bins / 2; ++k) {
+  for (unsigned long k = 0; k < bins / 2; ++k) {
 
     complex<double> sum;
-    for (unsigned int n = 0; n < bins; ++n)
+    for (unsigned long n = 0; n < bins; ++n)
       sum += twiddle[n][k] * complex<double>(samples.at(n), 0);
 
     // Store the absolute value of the complex average
